#100 feat: ycc 작품 총평 DB 저장 로직 구현

+++ src/main/java/com/susanghan_guys/server/personalwork/dto/response/YccWorkEvaluationResponse.java
@@ -4,15 +4,15 @@ import lombok.Builder;
 
 @Builder
 public record YccWorkEvaluationResponse(
-        Integer feasibilityScore,
+        double feasibilityScore,
         String feasibility,
-        Integer mediaScore,
+        double mediaScore,
         String media,
-        Integer agendaScore,
+        double agendaScore,
         String agenda,
-        Integer influenceScore,
+        double influenceScore,
         String influence,
-        Integer deliveryScore,
+        double deliveryScore,
         String delivery
 ) {
 }
